Move to `__thread` subfolder
