+++ cudax/include/cuda/experimental/__async/stop_token.cuh
@@ -21,7 +21,7 @@
 #  pragma system_header
 #endif // no system header
 
-#include <cuda/std/__atomic/wait/threading_support.h>
+#include <cuda/std/__thread/threading_support.h>
 #include <cuda/std/__type_traits/is_nothrow_constructible.h>
 #include <cuda/std/atomic>
