+++ libcudacxx/include/cuda/std/__thread/threading_support_cuda.h
@@ -7,8 +7,8 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef _LIBCUDACXX___ATOMIC_WAIT_THREADING_SUPPORT_CUDA_H
-#define _LIBCUDACXX___ATOMIC_WAIT_THREADING_SUPPORT_CUDA_H
+#ifndef _LIBCUDACXX___THREAD_THREADING_SUPPORT_CUDA_H
+#define _LIBCUDACXX___THREAD_THREADING_SUPPORT_CUDA_H
 
 #include <cuda/std/detail/__config>
 
@@ -45,4 +45,4 @@ _CCCL_POP_MACROS
 
 #endif // !_LIBCUDACXX_HAS_NO_THREADS && _LIBCUDACXX_HAS_THREAD_API_CUDA
 
-#endif // _LIBCUDACXX___ATOMIC_WAIT_THREADING_SUPPORT_CUDA_H
+#endif // _LIBCUDACXX___THREAD_THREADING_SUPPORT_CUDA_H
