+++ libcudacxx/include/cuda/std/__thread/threading_support_pthread.h
@@ -7,8 +7,8 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef _LIBCUDACXX___ATOMIC_WAIT_THREADING_SUPPORT_PTHREAD_H
-#define _LIBCUDACXX___ATOMIC_WAIT_THREADING_SUPPORT_PTHREAD_H
+#ifndef _LIBCUDACXX___THREAD_THREADING_SUPPORT_PTHREAD_H
+#define _LIBCUDACXX___THREAD_THREADING_SUPPORT_PTHREAD_H
 
 #include <cuda/std/detail/__config>
 
@@ -181,4 +181,4 @@ _CCCL_POP_MACROS
 
 #endif // !_LIBCUDACXX_HAS_NO_THREADS
 
-#endif // _LIBCUDACXX___ATOMIC_WAIT_THREADING_SUPPORT_PTHREAD_H
+#endif // _LIBCUDACXX___THREAD_THREADING_SUPPORT_PTHREAD_H
