+++ libcudacxx/include/cuda/std/__thread/threading_support_win32.h
@@ -7,8 +7,8 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef _LIBCUDACXX___ATOMIC_WAIT_THREADING_SUPPORT_WIN32_H
-#define _LIBCUDACXX___ATOMIC_WAIT_THREADING_SUPPORT_WIN32_H
+#ifndef _LIBCUDACXX___THREAD_THREADING_SUPPORT_WIN32_H
+#define _LIBCUDACXX___THREAD_THREADING_SUPPORT_WIN32_H
 
 #include <cuda/std/detail/__config>
 
@@ -86,4 +86,4 @@ _CCCL_POP_MACROS
 
 #endif // !_LIBCUDACXX_HAS_NO_THREADS && _LIBCUDACXX_HAS_THREAD_API_WIN32
 
-#endif // _LIBCUDACXX___ATOMIC_WAIT_THREADING_SUPPORT_H
+#endif // _LIBCUDACXX___THREAD_THREADING_SUPPORT_H
